Removed compiler warnings
git-svn-id: file:///media/bill/diskAshur2/backup/svn/microemacs/trunk@499 a259ec66-be10-11ea-b557-b34f8aa0b108

+++ nroff/nroff/dtags.c
@@ -20,7 +20,7 @@ tagAlloc (char *name, char *section, char *desc, char *file,
     if ((np = (TagP) malloc (sizeof (Tag))) == NULL)
         uFatal ("No memory\n");
 
-    if (isupper (name[0]))
+    if (isupper ((int)(name[0])))
         c = tolower (name[0]);
     else
         c = name[0];
